refactor(dt-ai-ingest): flatten utils/, consolidate dev deps, drop python-dotenv
