+++ dt-ai-ingest/pyproject.toml
@@ -34,14 +34,6 @@ dependencies = [
 parquet = [
     "pyarrow>=14",
 ]
-dev = [
-    "pytest>=7.0.0",
-    "pytest-asyncio>=1.3.0",
-    "pytest-cov>=4.0.0",
-    "pytest-httpx>=0.36.0",
-    "ruff>=0.1.0",
-    "mypy>=1.0.0",
-]
 
 [project.urls]
 Homepage = "https://github.com/dynatrace-oss/dt-ai-ingest"
@@ -80,7 +72,10 @@ ignore_missing_imports = true
 
 [dependency-groups]
 dev = [
+    "mypy>=1.0.0",
+    "pytest>=7.0.0",
     "pytest-asyncio>=1.3.0",
+    "pytest-cov>=4.0.0",
     "pytest-httpx>=0.36.0",
-    "python-dotenv>=1.2.2",
+    "ruff>=0.1.0",
 ]
