+++ dt-ai-ingest/src/dt_ai_ingest/spans.py
