--- dt-ai-ingest/src/dt_ai_ingest/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Small, pure helpers with no cross-module dependencies."""
